requirements.txt
pandas==1.2.4
pydeck==0.6.2
python-dateutil==2.8.1
streamlit==0.85.0